feat User Location
When application is opened, the map is centered on the user location by default. If location is not authorised/not available, camera is centered on the last map state saved. If no map state is saved, the default location is on Atlanta.
The map is saved when the fragment is stopped.
We look for the user location when the main activity is launched.